STAC-14481 moved container runtime interface out of util package.
