+++ pkg/collector/corechecks/container_runtime/container_if.go
@@ -1,4 +1,4 @@
-package util
+package container_runtime
 
 import (
 	"github.com/opencontainers/runtime-spec/specs-go"
